fix : login, logout and refactoring authenticate package
