+++ src/main/java/spaland/config/WebConfig.java
@@ -2,19 +2,19 @@ package spaland.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@EnableWebMvc
 public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
-                .allowCredentials(true)
-                .allowedOrigins("http://localhost:3000")
-                .allowedMethods("*")
-                .allowedOriginPatterns("*")
-        
-
-        ;
+                .allowCredentials(true).maxAge(3600)
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
+                .allowedHeaders("*")
+                .exposedHeaders("*")
+                .allowedOrigins("http://localhost:3000","http://127.0.0.1:3000");
     }
 }
