+++ src/main/java/spaland/users/service/IUserService.java
@@ -1,11 +1,12 @@
 package spaland.users.service;
 
-import spaland.users.vo.RequestUser;
+import spaland.users.model.User;
 import spaland.users.vo.ResponseUser;
+import spaland.users.vo.SignupRequest;
 
 public interface IUserService {
 
-    ResponseUser addUser(RequestUser requestUser);
+    User singup(SignupRequest signupRequest);
     ResponseUser getUser(Long id);
 
     boolean checkDuplicateId(String userId);
