+++ src/main/java/spaland/users/vo/LogoutResponse.java
@@ -1,17 +1,13 @@
-package spaland.auth;
+package spaland.users.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
-
-    private String userEmail;
-    private String password;
-
+public class LogoutResponse {
+    private String userNickname;
 }
