+++ src/main/java/spaland/users/vo/SignupRequest.java
@@ -1,14 +1,21 @@
 package spaland.users.vo;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
-public class RequestUser {
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class SignupRequest {
 
     private String userNickname;
     private String userEmail;
     private String userName;
     private String password;
     private String phone;
+
 
 }
