+++ src/main/java/spaland/utility/CookieUtil.java
@@ -1,4 +1,4 @@
-package spaland.auth;
+package spaland.utility;
 
 import org.springframework.http.ResponseCookie;
 import org.springframework.stereotype.Service;
@@ -12,18 +12,18 @@ import static spaland.config.JwtService.COOKIE_NAME;
 
 @Service
 public class CookieUtil {
-//    public ResponseCookie createCookie(String cookieName, String value){
-//        ResponseCookie cookie = ResponseCookie.from(cookieName, value)
-//                .maxAge(JwtService.REFRESH_TOKEN_VALIDATION_SECOND)
-//                .path("/")
-//                .secure(true)
-//                .httpOnly(true)
-//                .sameSite("None")
-//                .build();
-//        return cookie;
-//
-//
-//    }
+    public ResponseCookie createCookie(String cookieName, String value){
+        ResponseCookie cookie = ResponseCookie.from(cookieName, value)
+                .maxAge(JwtService.REFRESH_TOKEN_VALIDATION_SECOND)
+                .path("/")
+                .secure(true)
+                .httpOnly(true)
+                .sameSite("None")
+                .build();
+        return cookie;
+
+
+    }
 
     public Cookie cookie(String cookieName, String value){
         Cookie myCookie = new Cookie(cookieName,value);
